requirements.txt
# Libraries required for the Vehicle Re-Identification System
torch
torchaudio
torchvision
pytorch-metric-learning
pandas
tqdm

# From now on, the libraries are needed for the Tracking System
cycler==0.11.0
deep-sort-realtime==1.3.2
kiwisolver==1.4.4
markupsafe==2.1.2
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pyyaml==6.0
pillow>=10.3.0
requests>=2.32.0
setuptools>=70.0.0
ultralytics
opencv-python
opencv-contrib-python